# perf: Improve dashboard performance by decreasing rerenders