--- a/superset-frontend/src/dashboard/components/SyncDashboardState/index.tsx
+++ b/superset-frontend/src/dashboard/components/SyncDashboardState/index.tsx
@@ -53,7 +53,7 @@ const updateDashboardTabLocalStorage = (
   const dashboardsContexts = getDashboardContextLocalStorage();
   setItem(LocalStorageKeys.DashboardExploreContext, {
     ...dashboardsContexts,
-    [dashboardPageId]: dashboardContext,
+    [dashboardPageId]: { ...dashboardContext, dashboardPageId },
   });
 };
 
@@ -62,7 +62,7 @@ const selectDashboardContextForExplore = createSelector(
     (state: RootState) => state.dashboardInfo.metadata,
     (state: RootState) => state.dashboardInfo.id,
     (state: RootState) => state.dashboardState?.colorScheme,
-    (state: RootState) => state.nativeFilters.filters,
+    (state: RootState) => state.nativeFilters?.filters,
     (state: RootState) => state.dataMask,
   ],
   (metadata, dashboardId, colorScheme, filters, dataMask) => {
@@ -79,7 +79,7 @@ const selectDashboardContextForExplore = createSelector(
         metadata?.shared_label_colors,
       ),
       colorScheme,
-      chartConfiguration: metadata.chart_configuration || EMPTY_OBJECT,
+      chartConfiguration: metadata?.chart_configuration || EMPTY_OBJECT,
       nativeFilters: optimizedNativeFilters,
       dataMask,
       dashboardId,