--- a/superset-frontend/src/dashboard/components/gridComponents/Chart.jsx
+++ b/superset-frontend/src/dashboard/components/gridComponents/Chart.jsx
@@ -204,6 +204,10 @@ export const Chart = props => {
     };
   }, []);
 
+  useEffect(() => {
+    resize();
+  }, [props.isFullSize, props.height, props.width]);
+
   const getHeaderHeight = useCallback(() => {
     if (headerRef.current) {
       const computedStyle = getComputedStyle(