### SUMMARY
SUPERSEDED BY PRs: https://github.com/apache/superset/pull/31240, https://github.com/apache/superset/pull/31241, https://github.com/apache/superset/pull/31242, https://github.com/apache/superset/pull/31243, https://github.com/apache/superset/pull/31244.
This PR is not to be merged - aggregates 5 smaller PRs to be referenced as total performance improvement.


Due to many inefficient redux selectors and other factors causing numerous rerenders, large dashboards were running very slow.

This PR refactors many components related to Dashboard and native filters - some components are rewritten to functional to utilize React's memoization, Redux selectors optimized to select smaller objects that change less frequently, optimized frequently called functions.


### BEFORE/AFTER SCREENSHOTS OR ANIMATED GIF

The tests were conducted on a large dashboard - 300+ charts, 40+ filters. Test steps: reload the page, let the 1st tab load, switch to a different tab, let everything load.
Tested on my M2 Pro Macbook - since it's likely that an average Superset user works on a slower computer, I also run the tests with 4x and 20x CPU throttling.
Testing tool - chrome dev tools, performance tab

1. No throttling
Before - total scripting time 2.73s (screenshot 1), each chart holder blocks the main thread for ~70ms
After - total scripting time 1.48s (screenshot 2), each chart holder blocks the main thread for ~5ms

2. 4x throttling
Before - total scripting time 11.93s (screenshot 3), each chart holder blocks the main thread for ~300ms
After - total scripting time 6.23s (screenshot 4), each chart holder blocks the main thread for ~20ms

3. 20x throttling
Before - total scripting time 82s (screenshot 5), each chart holder blocks the main thread for ~2s
After - total scripting time 32s (screenshot 6), each chart holder blocks the main thread for ~100ms


**SCREENSHOTS BEFORE:**

1. No throttling

![image](https://github.com/user-attachments/assets/893043e5-f0a5-4e78-8fbb-e786c751abad)

2. 4x throttling

![image](https://github.com/user-attachments/assets/e84830a7-1f19-49f1-bd7f-aafdb099f376)

3. 20x throttling

![image](https://github.com/user-attachments/assets/f4bd2efe-ebbe-4134-ba51-bf9c861ecd0d)


**SCREENSHOTS AFTER:**

1. No throttling

![image](https://github.com/user-attachments/assets/fe695d7a-db0b-4c18-a2a1-5d618b99ec9a)

2. 4x throttling

![image](https://github.com/user-attachments/assets/5f62cceb-e329-4d1a-a9c7-9f00c0592092)

3. 20x throttling

![image](https://github.com/user-attachments/assets/011320c9-e321-4c90-b95b-0da5b27643d2)

### TESTING INSTRUCTIONS

No functional changes - all dashboard functionalities should work like before.

Recommend thorough testing of:
- Filter interactions
- Chart loading/rendering
- Dashboard navigation
- Cross-filter functionality
- Dashboard state persistence

### ADDITIONAL INFORMATION
<!--- Check any relevant boxes with "x" -->
<!--- HINT: Include "Fixes #nnn" if you are fixing an existing issue -->
- [ ] Has associated issue:
- [ ] Required feature flags:
- [ ] Changes UI
- [ ] Includes DB Migration (follow approval process in [SIP-59](https://github.com/apache/superset/issues/13351))
  - [ ] Migration is atomic, supports rollback & is backwards-compatible
  - [ ] Confirm DB migration upgrade and downgrade tested
  - [ ] Runtime estimates and downtime expectations provided
- [ ] Introduces new feature or API
- [ ] Removes existing feature or API