@villebro 2024-11-18
/testenv up